feat(cli): refactor v0.1.2 and add show-mnemonic
